port to D language version 2
With this patch the code is accepted by gdc-4.6 0.29.1-4.6.4-3 without
triggering deprecated feature warnings/errors.

+++ src/abagames/util/ActorPool.d
@@ -16,7 +16,7 @@ public class ActorPool {
  public:
   Actor[] actor;
  protected:
-  int actorIdx;
+  ptrdiff_t actorIdx;
 
   public this(int n, Actor act, ActorInitializer ini) {
     actor = new Actor[n];
